requirements.txt
APScheduler==3.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gensim==3.8.3
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
newsapi-python==0.2.6
pandas==1.4.3
pycparser==2.20
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
WTForms==2.3.3
